+++ stubs/inertia-react/resources/js/Components/Dropdown/DropdownContent.jsx
@@ -1,6 +1,6 @@
 import { Transition } from '@headlessui/react';
 import { Fragment } from 'react';
-import { useDropdownContext } from '../../Contexts/DropdownContext'
+import { useDropdownContext } from './DropdownContext'
 
 
 export default function DropdownContent({ align = 'right', width = '48', contentClasses = 'py-1 bg-white dark:bg-gray-700', children }) {
