+++ stubs/inertia-react/resources/js/Components/Dropdown/DropdownTrigger.jsx
@@ -1,4 +1,4 @@
-import { useDropdownContext } from '../../Contexts/DropdownContext'
+import { useDropdownContext } from './DropdownContext'
 
 export default function DropdownTrigger({ children }) {
     const { isOpen, setIsOpen, handleToggleOpen } = useDropdownContext()
